backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
strands-agents==0.1.5
strands-agents-tools==0.1.2
boto3==1.36.26
python-multipart==0.0.20
